Rename :ui:home to :ui:userlist, extract user list routes to UserListRoutes object
