+++ ui/userlist/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="title">JetpackComposeTemplate</string>
+    <string name="title">User list</string>
 </resources>
